+++ jsonrpc/chain.go
@@ -2,6 +2,9 @@ package jsonrpc
 
 import (
 	"encoding/hex"
+	"fmt"
+	"github.com/ngchain/ngcore/ngstate"
+	"github.com/ngchain/ngcore/ngtypes"
 
 	jsonrpc2 "github.com/maoxs2/go-jsonrpc2"
 	"github.com/ngchain/ngcore/storage"
@@ -65,6 +68,11 @@ type getTxByHashParams struct {
 	Hash string `json:"hash"`
 }
 
+type getTxByHashReply struct {
+	OnChain bool `json:"onChain"`
+	Tx *ngtypes.Tx `json:"tx"`
+}
+
 func (s *Server) getTxByHashFunc(msg *jsonrpc2.JsonRpcMessage) *jsonrpc2.JsonRpcMessage {
 	var params getTxByHashParams
 	err := utils.JSON.Unmarshal(msg.Params, params)
@@ -82,10 +90,32 @@ func (s *Server) getTxByHashFunc(msg *jsonrpc2.JsonRpcMessage) *jsonrpc2.JsonRpc
 		return jsonrpc2.NewJsonRpcError(msg.ID, jsonrpc2.NewError(0, err))
 	}
 
-	raw, err := utils.JSON.Marshal(tx)
-	if err != nil {
-		return jsonrpc2.NewJsonRpcError(msg.ID, jsonrpc2.NewError(0, err))
+	if tx != nil {
+		raw, err := utils.JSON.Marshal(&getTxByHashReply{
+			OnChain: true,
+			Tx: tx,
+		})
+		if err != nil {
+			return jsonrpc2.NewJsonRpcError(msg.ID, jsonrpc2.NewError(0, err))
+		}
+
+		return jsonrpc2.NewJsonRpcSuccess(msg.ID, raw)
 	}
 
-	return jsonrpc2.NewJsonRpcSuccess(msg.ID, raw)
+	// search in pool
+	pool := ngstate.GetTxPool()
+	exists, tx := pool.IsInPool(hash)
+	if exists && tx != nil {
+		raw, err := utils.JSON.Marshal(&getTxByHashReply{
+			OnChain: true,
+			Tx: tx,
+		})
+		if err != nil {
+			return jsonrpc2.NewJsonRpcError(msg.ID, jsonrpc2.NewError(0, err))
+		}
+
+		return jsonrpc2.NewJsonRpcSuccess(msg.ID, raw)
+	}
+
+	return jsonrpc2.NewJsonRpcError(msg.ID, jsonrpc2.NewError(0, fmt.Errorf("cannot find the tx with hash %x", hash)))
 }